Migrated registration to register.js
