+++ views/register.ejs
@@ -2,8 +2,8 @@
 <html lang="en">
 
 <head>
-  <!-- Including Top Headers -->
-  <% include ../views/partials/topHeaders.ejs %>
+    <!-- Including Top Headers -->
+    <% include ../views/partials/topHeaders.ejs %>
 </head>
 
 <body>
@@ -11,44 +11,63 @@
     <% include ../views/partials/topNav.ejs %>
 
     <!-- Main Content Container -->
-  <div class="container">
-    <div class="row">
-      <div class="col">
-        <form method="post">
-          <div class="form-group">
-            <label for="exampleInputEmail1">First Name</label>
-            <input type="text" class="form-control" id="firstName" placeholder="First Name">
-          </div>
-          <div class="form-group">
-            <label for="exampleInputEmail1">Last Name</label>
-            <input type="text" class="form-control" id="lastName" placeholder="Last Name">
-          </div>
-          <div class="form-group">
-            <label for="exampleInputEmail1">Email</label>
-            <input type="email" class="form-control" id="email" placeholder="Email">
-          </div>
-          <div class="form-group">
-            <label for="exampleInputEmail1">Username</label>
-            <input type="text" class="form-control" id="user" placeholder="Username">
-          </div>
-          <div class="form-group">
-            <label for="exampleInputPassword1">Password</label>
-            <input type="password" class="form-control" id="pass" placeholder="Password">
-          </div>
-          <div class="form-group">
-            <label for="exampleInputPassword1">Confirm Password</label>
-            <input type="password" class="form-control" id="confPass" placeholder="Confirm Password">
-          </div>
-          <button type="submit" class="btn btn-primary">Register</button>
+    <div class="container">
+        <p class="error"><%=error%></p>
+        <form method="post" action="/register">
+            <div class="row">
+                <div class="col">
+                    <div class="form-group">
+                        <label for="exampleInputEmail1">First Name</label>
+                        <input type="text" class="form-control" name="firstName" placeholder="First Name">
+                    </div>
+                    <div class="form-group">
+                        <label for="exampleInputEmail1">Last Name</label>
+                        <input type="text" class="form-control" name="lastName" placeholder="Last Name">
+                    </div>
+                    <div class="form-group">
+                        <label for="exampleInputEmail1">Email</label>
+                        <input type="email" class="form-control" name="email" placeholder="Email">
+                    </div>
+                    <div class="form-group">
+                        <label for="exampleInputEmail1">Username</label>
+                        <input type="text" class="form-control" name="user" placeholder="Username">
+                    </div>
+                    <div class="form-group">
+                        <label for="exampleInputPassword1">Password</label>
+                        <input type="password" class="form-control" name="pass" placeholder="Password">
+                    </div>
+                    <div class="form-group">
+                        <label for="exampleInputPassword1">Confirm Password</label>
+                        <input type="password" class="form-control" name="confPass" placeholder="Confirm Password">
+                    </div>
+                    <button type="submit" class="btn btn-primary">Register</button>
+                </div>
+                <div class="col">
+                        <label for="year">Year</label>
+                        <select name="year" class="form-control">
+                                <option>Freshman</option>
+                                <option>Sophomore</option>
+                                <option>Junior</option>
+                                <option>Senior</option>
+                                <option>Grad Student</option>
+                        </select>
+                        <label for="AOS">Area of Study</label>
+                        <select name="aos" class="form-control">
+                                <option>Web Dev</option>
+                                <option>Robotics</option>
+                                <option>Security</option>
+                                <option>Networking</option>
+                                <option>Not Sure</option>
+                        </select>
+                </div>
+            </div>
         </form>
-      </div>
     </div>
-  </div>
-  <!-- Including Footer -->
-  <% include ../views/partials/botHeaders.ejs %>
+    <!-- Including Footer -->
+    <% include ../views/partials/botHeaders.ejs %>
 
-  <!-- Including bottom headers -->
-  <% include ../views/partials/botHeaders.ejs %>
+    <!-- Including bottom headers -->
+    <% include ../views/partials/botHeaders.ejs %>
 </body>
 
 </html>
